# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pdomain-prep-for-pgdp",
    "description": "Convert a folder/zip of scanned book images into a PGDP-ready submission package.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/data/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Project",
        "operationId": "delete_project",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/gpu/ingest": {
      "post": {
        "tags": [
          "gpu"
        ],
        "summary": "Ingest",
        "operationId": "start_ingest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/gpu/suggest-splits": {
      "post": {
        "tags": [
          "gpu"
        ],
        "summary": "Suggest Splits",
        "operationId": "suggest_splits",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestSplitsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestSplitsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/gpu/suggest-illustrations": {
      "post": {
        "tags": [
          "gpu"
        ],
        "summary": "Suggest Illustrations",
        "operationId": "suggest_illustrations",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestIllustrationsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestIllustrationsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/gpu/extract-illustration": {
      "post": {
        "tags": [
          "gpu"
        ],
        "summary": "Extract Illustration",
        "operationId": "extract_illustration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractIllustrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractIllustrationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlignmentOverride": {
        "type": "string",
        "enum": [
          "default",
          "top",
          "center",
          "bottom"
        ],
        "title": "AlignmentOverride"
      },
...
      "ExtractIllustrationRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "idx0": {
            "type": "integer",
            "title": "Idx0"
          },
          "region_index": {
            "type": "integer",
            "title": "Region Index"
          },
          "output_format": {
            "type": "string",
            "enum": [
              "jpg",
              "png"
            ],
            "title": "Output Format",
            "default": "jpg"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "idx0",
          "region_index"
        ],
        "title": "ExtractIllustrationRequest"
      },
      "ExtractIllustrationResponse": {
        "properties": {
          "image_key": {
            "type": "string",
            "title": "Image Key"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          }
        },
        "type": "object",
        "required": [
          "image_key",
          "image_url"
        ],
        "title": "ExtractIllustrationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IllustrationRegion-Output": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index",
            "default": 1
          },
          "label": {
            "type": "string",
            "title": "Label",
            "default": ""
          },
          "type": {
            "type": "string",
            "enum": [
              "illustration",
              "decoration",
              "plate"
            ],
            "title": "Type",
            "default": "illustration"
          },
          "L": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "L"
          },
          "R": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "R"
          },
          "T": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "T"
          },
          "B": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "B"
          },
          "output_format": {
            "type": "string",
            "enum": [
              "jpg",
              "png"
            ],
            "title": "Output Format",
            "default": "jpg"
          },
          "jpeg_quality": {
            "type": "integer",
            "title": "Jpeg Quality",
            "default": 85
          },
          "convert_to_grayscale": {
            "type": "boolean",
            "title": "Convert To Grayscale",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "index",
          "label",
          "type",
          "L",
          "R",
          "T",
          "B",
          "output_format",
          "jpeg_quality",
          "convert_to_grayscale"
        ],
        "title": "IllustrationRegion",
        "description": "Coords in SOURCE image space (original scan, pre-processing)."
      },
      "IngestRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "source_key": {
            "type": "string",
            "title": "Source Key"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "zip",
              "s3_folder",
              "local_folder"
            ],
            "title": "Source Type"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "source_key",
          "source_type"
        ],
        "title": "IngestRequest"
      },
...
      "JobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running"
            ],
            "title": "Status",
            "default": "queued"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "JobResponse"
      },
...
      "PageSplit-Output": {
        "properties": {
          "suffix": {
            "type": "string",
            "title": "Suffix"
          },
          "reading_order": {
            "type": "integer",
            "title": "Reading Order"
          },
          "L": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "L"
          },
          "R": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "R"
          },
          "T": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "T"
          },
          "B": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "B"
          },
          "scale_to_standard_page": {
            "type": "boolean",
            "title": "Scale To Standard Page",
            "default": true
          },
          "alignment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlignmentOverride"
              },
              {
                "type": "null"
              }
            ]
          },
          "ocr_engine": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "doctr",
                  "tesseract"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Engine"
          }
        },
        "type": "object",
        "required": [
          "suffix",
          "reading_order",
          "L",
          "R",
          "T",
          "B",
          "scale_to_standard_page",
          "alignment",
          "ocr_engine"
        ],
        "title": "PageSplit",
        "description": "Replaces the notebook's `PageSectionSplit`. Coords in PROCESSED image space."
      },
...
      "PipelineState": {
        "properties": {
          "steps": {
            "additionalProperties": {
              "$ref": "#/components/schemas/StepState"
            },
            "type": "object",
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "steps"
        ],
        "title": "PipelineState"
      },
      "Project": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id",
            "default": "default"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "proof_page_count": {
            "type": "integer",
            "title": "Proof Page Count"
          },
          "config": {
            "$ref": "#/components/schemas/ProjectConfig"
          },
          "pipeline_state": {
            "$ref": "#/components/schemas/PipelineState"
          },
          "storage_prefix": {
            "type": "string",
            "title": "Storage Prefix"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived",
            "default": false
          },
          "stage_artifacts_bytes": {
            "type": "integer",
            "title": "Stage Artifacts Bytes",
            "default": 0
          },
          "source_zip_bytes": {
            "type": "integer",
            "title": "Source Zip Bytes",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "name",
          "created_at",
          "updated_at",
          "status",
          "page_count",
          "proof_page_count",
          "config",
          "pipeline_state",
          "storage_prefix",
          "archived",
          "stage_artifacts_bytes",
          "source_zip_bytes"
        ],
        "title": "Project"
      },
      "ProjectConfig": {
        "properties": {
          "book_name": {
            "type": "string",
            "title": "Book Name"
          },
          "source_uri": {
            "type": "string",
            "title": "Source Uri"
          },
          "proof_start_idx0": {
            "type": "integer",
            "title": "Proof Start Idx0",
            "default": 0
          },
          "proof_end_idx0": {
            "type": "integer",
            "title": "Proof End Idx0",
            "default": 0
          },
          "cover_idx0": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Idx0"
          },
          "title_idx0": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Idx0"
          },
          "frontmatter_start_idx0": {
            "type": "integer",
            "title": "Frontmatter Start Idx0",
            "default": 0
          },
          "frontmatter_end_idx0": {
            "type": "integer",
            "title": "Frontmatter End Idx0",
            "default": 0
          },
          "bodymatter_start_idx0": {
            "type": "integer",
            "title": "Bodymatter Start Idx0",
            "default": 0
          },
          "bodymatter_end_idx0": {
            "type": "integer",
            "title": "Bodymatter End Idx0",
            "default": 0
          },
          "frontmatter_page_nbr_start": {
            "type": "integer",
            "title": "Frontmatter Page Nbr Start",
            "default": 1
          },
          "bodymatter_page_nbr_start": {
            "type": "integer",
            "title": "Bodymatter Page Nbr Start",
            "default": 1
          },
          "initial_crop_all": {
            "prefixItems": [
              {
                "type": "integer"
              },
              {
                "type": "integer"
              },
              {
                "type": "integer"
              },
              {
                "type": "integer"
              }
            ],
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Initial Crop All",
            "default": [
              0,
              0,
              0,
              0
            ]
          },
          "ocr_crop_top": {
            "type": "integer",
            "title": "Ocr Crop Top",
            "default": 0
          },
          "ocr_crop_bottom": {
            "type": "integer",
            "title": "Ocr Crop Bottom",
            "default": 0
          },
          "ocr_crop_left": {
            "type": "integer",
            "title": "Ocr Crop Left",
            "default": 0
          },
          "ocr_crop_right": {
            "type": "integer",
            "title": "Ocr Crop Right",
            "default": 0
          },
          "custom_regex_passes": {
            "items": {
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Custom Regex Passes"
          },
          "custom_scannos": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Custom Scannos"
          },
          "layout_category_overrides": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Layout Category Overrides"
          },
          "optimize_png": {
            "type": "boolean",
            "title": "Optimize Png",
            "default": true
          },
          "default_overrides": {
            "additionalProperties": true,
            "type": "object",
            "title": "Default Overrides"
          }
        },
        "type": "object",
        "required": [
          "book_name",
          "source_uri",
          "proof_start_idx0",
          "proof_end_idx0",
          "cover_idx0",
          "title_idx0",
          "frontmatter_start_idx0",
          "frontmatter_end_idx0",
          "bodymatter_start_idx0",
          "bodymatter_end_idx0",
          "frontmatter_page_nbr_start",
          "bodymatter_page_nbr_start",
          "initial_crop_all",
          "ocr_crop_top",
          "ocr_crop_bottom",
          "ocr_crop_left",
          "ocr_crop_right",
          "custom_regex_passes",
          "custom_scannos",
          "layout_category_overrides",
          "optimize_png",
          "default_overrides"
        ],
        "title": "ProjectConfig"
      },
      "ProjectStatus": {
        "type": "string",
        "enum": [
          "ingesting",
          "configuring",
          "processing",
          "reviewing",
          "packaging",
          "complete"
        ],
        "title": "ProjectStatus"
      },
...
      "StepState": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/StepStatus",
            "default": "pending"
          },
          "pages_complete": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Pages Complete"
          },
          "pages_error": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Pages Error"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "pages_complete",
          "pages_error",
          "started_at",
          "completed_at",
          "job_id"
        ],
        "title": "StepState"
      },
      "StepStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "complete",
          "error"
        ],
        "title": "StepStatus"
      },
      "SuggestIllustrationsRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "idx0": {
            "type": "integer",
            "title": "Idx0"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "idx0"
        ],
        "title": "SuggestIllustrationsRequest"
      },
      "SuggestIllustrationsResponse": {
        "properties": {
          "regions": {
            "items": {
              "$ref": "#/components/schemas/IllustrationRegion-Output"
            },
            "type": "array",
            "title": "Regions",
            "default": []
          }
        },
        "type": "object",
        "title": "SuggestIllustrationsResponse"
      },
      "SuggestSplitsRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "idx0": {
            "type": "integer",
            "title": "Idx0"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "idx0"
        ],
        "title": "SuggestSplitsRequest"
      },
      "SuggestSplitsResponse": {
        "properties": {
          "splits": {
            "items": {
              "$ref": "#/components/schemas/PageSplit-Output"
            },
            "type": "array",
            "title": "Splits",
            "default": []
          }
        },
        "type": "object",
        "title": "SuggestSplitsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}